[FIRRTL][HW] Move InnerRef Infra to HW
Move the InnerRef infrastructure from the FIRRTL Dialect to the HW
Dialect.  This is done in preparation to move the HierPathOp to the HW
Dialect (which is further preparation for replacing the GlobalRef op
with this).
