+++ include/circt/Dialect/FIRRTL/FIRRTLOpInterfaces.h
@@ -15,17 +15,15 @@
 #define CIRCT_DIALECT_FIRRTL_OP_INTERFACES_H
 
 #include "circt/Dialect/FIRRTL/FIRRTLAnnotations.h"
-#include "circt/Dialect/FIRRTL/FIRRTLAttributes.h"
 #include "circt/Dialect/FIRRTL/FIRRTLTypes.h"
-#include "circt/Dialect/FIRRTL/InnerSymbolTable.h"
 #include "circt/Dialect/HW/HWAttributes.h"
 #include "circt/Dialect/HW/HWOpInterfaces.h"
+#include "circt/Dialect/HW/InnerSymbolTable.h"
 #include "mlir/IR/Attributes.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/OpDefinition.h"
 #include "mlir/IR/SymbolTable.h"
 #include "llvm/ADT/TypeSwitch.h"
-#include "llvm/Support/CommandLine.h"
 
 namespace circt {
 namespace firrtl {
@@ -100,62 +98,8 @@ struct PortInfo {
 /// Verification hook for verifying module like operations.
 LogicalResult verifyModuleLikeOpInterface(FModuleLike module);
 
-class InnerSymbolOpInterface;
-/// Verification hook for verifying InnerSym Attribute.
-LogicalResult verifyInnerSymAttr(InnerSymbolOpInterface op);
-
-namespace detail {
-LogicalResult verifyInnerRefNamespace(Operation *op);
-} // namespace detail
-
 } // namespace firrtl
 } // namespace circt
-
-namespace mlir {
-namespace OpTrait {
-
-/// This trait is for operations that define a scope for resolving InnerRef's,
-/// and provides verification for InnerRef users (via InnerRefUserOpInterface).
-template <typename ConcreteType>
-class InnerRefNamespace : public TraitBase<ConcreteType, InnerRefNamespace> {
-public:
-  static LogicalResult verifyRegionTrait(Operation *op) {
-    static_assert(
-        ConcreteType::template hasTrait<::mlir::OpTrait::SymbolTable>(),
-        "expected operation to be a SymbolTable");
-
-    if (op->getNumRegions() != 1)
-      return op->emitError("expected operation to have a single region");
-    if (!op->getRegion(0).hasOneBlock())
-      return op->emitError("expected operation to have a single block");
-
-    // Verify all InnerSymbolTable's and InnerRef users.
-    return ::circt::firrtl::detail::verifyInnerRefNamespace(op);
-  }
-};
-
-/// A trait for inner symbol table functionality on an operation.
-template <typename ConcreteType>
-class InnerSymbolTable : public TraitBase<ConcreteType, InnerSymbolTable> {
-public:
-  static LogicalResult verifyRegionTrait(Operation *op) {
-    // Insist that ops with InnerSymbolTable's provide a Symbol, this is
-    // essential to how InnerRef's work.
-    static_assert(
-        ConcreteType::template hasTrait<::mlir::SymbolOpInterface::Trait>(),
-        "expected operation to define a Symbol");
-
-    // InnerSymbolTable's must be directly nested within an InnerRefNamespace.
-    auto *parent = op->getParentOp();
-    if (!parent || !parent->hasTrait<InnerRefNamespace>())
-      return op->emitError(
-          "InnerSymbolTable must have InnerRefNamespace parent");
-
-    return success();
-  }
-};
-} // namespace OpTrait
-} // namespace mlir
 
 #include "circt/Dialect/FIRRTL/FIRRTLOpInterfaces.h.inc"
 #endif // CIRCT_DIALECT_FIRRTL_OP_INTERFACES_H
