+++ include/circt/Dialect/FIRRTL/FIRRTLOpInterfaces.td
@@ -306,108 +306,6 @@ def FConnectLike : OpInterface<"FConnectLike"> {
   ];
 }
 
-def InnerRefNamespace : NativeOpTrait<"InnerRefNamespace">;
-
-def InnerSymbol : OpInterface<"InnerSymbolOpInterface"> {
-  let description = [{
-    This interface describes an operation that may define an
-    `inner_sym`.  An `inner_sym` operation resides 
-    in arbitrarily-nested regions of a region that defines a
-    `InnerSymbolTable`.
-    Inner Symbols are different from normal symbols due to 
-    MLIR symbol table resolution rules.  Specifically normal
-    symbols are resolved by first going up to the closest
-    parent symbol table and resolving from there (recursing
-    down for complex symbol paths).  In FIRRTL and SV, modules
-    define a symbol in a circuit or std.module symbol table.
-    For instances to be able to resolve the modules they
-    instantiate, the symbol use in an instance must resolve 
-    in the top-level symbol table.  If a module were a
-    symbol table, instances resolving a symbol would start from 
-    their own module, never seeing other modules (since 
-    resolution would start in the parent module of the 
-    instance and be unable to go to the global scope).
-    The second problem arises from nesting.  Symbols defining 
-    ops must be immediate children of a symbol table.  FIRRTL
-    and SV operations which define a inner_sym are grandchildren,
-    at least, of a symbol table and may be much further nested.
-    Lastly, ports need to define inner_sym, something not allowed
-    by normal symbols.
-
-    Any operation implementing an InnerSymbol may have the inner symbol be
-    optional and all methods should be robuse to the attribute not being
-    defined.
-  }];
-
-  let cppNamespace = "::circt::firrtl";
-  let methods = [
-    InterfaceMethod<"Returns the name of this inner symbol.",
-      "::mlir::StringAttr", "getInnerNameAttr", (ins), [{}],
-      /*defaultImplementation=*/[{
-        if (auto attr =
-              this->getOperation()->template getAttrOfType<hw::InnerSymAttr>(
-              circt::firrtl::InnerSymbolTable::getInnerSymbolAttrName()))
-          return attr.getSymName();
-        return {};
-      }]
-    >,
-    InterfaceMethod<"Returns the name of this inner symbol.",
-      "Optional<::mlir::StringRef>", "getInnerName", (ins), [{}],
-      /*defaultImplementation=*/[{
-        auto attr = this->getInnerNameAttr();
-        return attr ? Optional<StringRef>(attr.getValue()) : std::nullopt;
-      }]
-    >,
-    InterfaceMethod<"Sets the name of this inner symbol.",
-      "void", "setInnerSymbol", (ins "::mlir::StringAttr":$name), [{}],
-      /*defaultImplementation=*/[{
-        this->getOperation()->setAttr(
-            InnerSymbolTable::getInnerSymbolAttrName(), hw::InnerSymAttr::get(name));
-      }]
-    >,
-    InterfaceMethod<"Returns an InnerRef to this operation.  Must have inner symbol.",
-      "::circt::hw::InnerRefAttr", "getInnerRef", (ins), [{}],
-      /*defaultImplementation=*/[{
-        auto *op = this->getOperation();
-        return hw::InnerRefAttr::get(
-            SymbolTable::getSymbolName(
-                op->template getParentWithTrait<OpTrait::InnerSymbolTable>()),
-            InnerSymbolTable::getInnerSymbol(op));
-      }]
-    >,
-    InterfaceMethod<"Returns the InnerSymAttr.",
-      "::circt::hw::InnerSymAttr", "getInnerSymAttr", (ins), [{}],
-      /*defaultImplementation=*/[{
-        return this->getOperation()->template getAttrOfType<hw::InnerSymAttr>(
-              circt::firrtl::InnerSymbolTable::getInnerSymbolAttrName());
-      }]
-    >,
-  ];
-
-  let verify = [{
-    return verifyInnerSymAttr(op);
-  }];
-}
-
-def InnerSymbolTable : NativeOpTrait<"InnerSymbolTable">;
-
-def InnerRefUserOpInterface : OpInterface<"InnerRefUserOpInterface"> {
-  let description = [{
-    This interface describes an operation that may use a `InnerRef`. This
-    interface allows for users of inner symbols to hook into verification and
-    other inner symbol related utilities that are either costly or otherwise
-    disallowed within a traditional operation.
-  }];
-  let cppNamespace = "::circt::firrtl";
-
-  let methods = [
-    InterfaceMethod<"Verify the inner ref uses held by this operation.",
-      "::mlir::LogicalResult", "verifyInnerRefs",
-      (ins "::circt::firrtl::InnerRefNamespace&":$ns)
-    >,
-  ];
-}
-
 def FNamableOp : OpInterface<"FNamableOp"> {
   let cppNamespace = "circt::firrtl";
   let description = [{The interface provides common methods for namable operations
