+++ include/circt/Dialect/FIRRTL/FIRRTLOps.h
@@ -18,6 +18,7 @@
 #include "circt/Dialect/HW/HWAttributes.h"
 #include "circt/Dialect/HW/HWOpInterfaces.h"
 #include "circt/Dialect/HW/HWTypes.h"
+#include "circt/Dialect/HW/InnerSymbolTable.h"
 #include "circt/Support/FieldRef.h"
 #include "mlir/IR/Builders.h"
 #include "mlir/IR/FunctionInterfaces.h"
@@ -98,7 +99,7 @@ inline MemDirAttr &operator|=(MemDirAttr &lhs, MemDirAttr rhs) {
 /// Return the StringAttr for the inner_sym name, if it exists.
 inline StringAttr getInnerSymName(Operation *op) {
   auto s = op->getAttrOfType<hw::InnerSymAttr>(
-      InnerSymbolTable::getInnerSymbolAttrName());
+      hw::InnerSymbolTable::getInnerSymbolAttrName());
   if (s)
     return s.getSymName();
   return StringAttr();
