+++ include/circt/Dialect/HW/HWOpInterfaces.h
@@ -13,6 +13,8 @@
 #ifndef CIRCT_DIALECT_HW_HWOPINTERFACES_H
 #define CIRCT_DIALECT_HW_HWOPINTERFACES_H
 
+#include "circt/Dialect/HW/InnerSymbolTable.h"
+#include "circt/Support/LLVM.h"
 #include "mlir/IR/OpDefinition.h"
 #include "mlir/IR/SymbolTable.h"
 
@@ -21,6 +23,66 @@ namespace hw {
 struct PortInfo;
 } // namespace hw
 } // namespace circt
+
+namespace circt {
+namespace hw {
+
+class InnerSymbolOpInterface;
+/// Verification hook for verifying InnerSym Attribute.
+LogicalResult verifyInnerSymAttr(InnerSymbolOpInterface op);
+
+namespace detail {
+LogicalResult verifyInnerRefNamespace(Operation *op);
+} // namespace detail
+
+} // namespace hw
+} // namespace circt
+
+namespace mlir {
+namespace OpTrait {
+
+/// This trait is for operations that define a scope for resolving InnerRef's,
+/// and provides verification for InnerRef users (via InnerRefUserOpInterface).
+template <typename ConcreteType>
+class InnerRefNamespace : public TraitBase<ConcreteType, InnerRefNamespace> {
+public:
+  static LogicalResult verifyRegionTrait(Operation *op) {
+    static_assert(
+        ConcreteType::template hasTrait<::mlir::OpTrait::SymbolTable>(),
+        "expected operation to be a SymbolTable");
+
+    if (op->getNumRegions() != 1)
+      return op->emitError("expected operation to have a single region");
+    if (!op->getRegion(0).hasOneBlock())
+      return op->emitError("expected operation to have a single block");
+
+    // Verify all InnerSymbolTable's and InnerRef users.
+    return ::circt::hw::detail::verifyInnerRefNamespace(op);
+  }
+};
+
+/// A trait for inner symbol table functionality on an operation.
+template <typename ConcreteType>
+class InnerSymbolTable : public TraitBase<ConcreteType, InnerSymbolTable> {
+public:
+  static LogicalResult verifyRegionTrait(Operation *op) {
+    // Insist that ops with InnerSymbolTable's provide a Symbol, this is
+    // essential to how InnerRef's work.
+    static_assert(
+        ConcreteType::template hasTrait<::mlir::SymbolOpInterface::Trait>(),
+        "expected operation to define a Symbol");
+
+    // InnerSymbolTable's must be directly nested within an InnerRefNamespace.
+    auto *parent = op->getParentOp();
+    if (!parent || !parent->hasTrait<InnerRefNamespace>())
+      return op->emitError(
+          "InnerSymbolTable must have InnerRefNamespace parent");
+
+    return success();
+  }
+};
+} // namespace OpTrait
+} // namespace mlir
 
 #include "circt/Dialect/HW/HWOpInterfaces.h.inc"
