+++ lib/Dialect/HW/CMakeLists.txt
@@ -11,18 +11,19 @@ add_circt_dialect_library(
   InstanceGraphBase.cpp
   InstanceImplementation.cpp
   ModuleImplementation.cpp
-  
+  InnerSymbolTable.cpp
+
   ADDITIONAL_HEADER_DIRS
   ${CIRCT_MAIN_INCLUDE_DIR}/circt/Dialect/HW
-  
+
   DEPENDS
   MLIRHWIncGen
   MLIRHWAttrIncGen
   MLIRHWEnumsIncGen
-  
+
   LINK_COMPONENTS
   Support
-  
+
   LINK_LIBS PUBLIC
   MLIRIR
 )
