+++ lib/Dialect/HW/HWOpInterfaces.cpp
@@ -10,9 +10,69 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "circt/Dialect/HW/HWOpInterfaces.h"
 #include "circt/Dialect/HW/HWOps.h"
+#include "circt/Support/LLVM.h"
 #include "mlir/IR/BuiltinAttributes.h"
 #include "mlir/IR/BuiltinTypes.h"
+#include "llvm/ADT/SmallBitVector.h"
+#include "llvm/ADT/SmallPtrSet.h"
 #include "llvm/ADT/StringRef.h"
+
+using namespace circt;
+
+LogicalResult hw::verifyInnerSymAttr(InnerSymbolOpInterface op) {
+  auto innerSym = op.getInnerSymAttr();
+  // If does not have any inner sym then ignore.
+  if (!innerSym)
+    return success();
+  if (isa<InstanceOp>(&op) || op->getNumResults() != 1) {
+    // The inner sym can only be specified on fieldID=0.
+    if (innerSym.size() > 1 || !innerSym.getSymName()) {
+      op->emitOpError("cannot assign symbols to non-zero field id, for ops "
+                      "with zero or multiple results");
+      return failure();
+    }
+    return success();
+  }
+
+  // ----------------------------------------
+  // auto resultType =
+  // op->getResultTypes()[0].dyn_cast<firrtl::FIRRTLBaseType>(); if
+  // (!resultType)
+  //   return op->emitOpError("cannot attach symbols to non-base types");
+  // auto maxFields = resultType.getMaxFieldID();
+  // llvm::SmallBitVector indices(maxFields + 1);
+  // llvm::SmallPtrSet<Attribute, 8> symNames;
+  // // Ensure fieldID and symbol names are unique.
+  // auto uniqSyms = [&](hw::InnerSymPropertiesAttr p) {
+  //   if (maxFields < p.getFieldID()) {
+  //     op->emitOpError("field id:'" + Twine(p.getFieldID()) +
+  //                     "' is greater than the maximum field id:'" +
+  //                     Twine(maxFields) + "'");
+  //     return false;
+  //   }
+  //   if (indices.test(p.getFieldID())) {
+  //     op->emitOpError("cannot assign multiple symbol names to the field id:'"
+  //     +
+  //                     Twine(p.getFieldID()) + "'");
+  //     return false;
+  //   }
+  //   indices.set(p.getFieldID());
+  //   auto it = symNames.insert(p.getName());
+  //   if (!it.second) {
+  //     op->emitOpError("cannot reuse symbol name:'" + p.getName().getValue() +
+  //                     "'");
+  //     return false;
+  //   }
+  //   return true;
+  // };
+
+  // if (!llvm::all_of(innerSym.getProps(), uniqSyms))
+  //   return failure();
+  // ----------------------------------------
+
+  return success();
+}
 
 #include "circt/Dialect/HW/HWOpInterfaces.cpp.inc"
